move ssoAssociation to pipe_tasks
